Update package template based on mumble_1.2.8-2build1.debian
This is pretty much copy and paste with minor adjustments for patches
we have already applied as well as some other maintenance.

Drops lenny, maverick and natty support as those are EOL. Lucid
desktop is EOL but currently still has server support. That ends
at the end of this month though so modifying it to only build server
doesn't seem worth it so we drop it too.

Adds backport files for trusty and utopic as they do not ship with
opus.