Make PPA use Qt 5 by default
Except 12.04 LTS all Ubuntu versions supported by us have Qt 5 packages.
To be able to cleanly switch to it for the next release this patch makes
it the default and adjusts the backport script for 12.04 LTS to use the
old qmake as well as the 'qt4-legacy-compat' flag.

Fixes issue #3